feat: client collection and features crud support
