--- ogcapi-client/src/client.rs
@@ -1,344 +1,0 @@
-use std::future::Future;
-use std::pin::Pin;
-use std::task::{Context, Poll};
-
-use futures_core::Stream;
-use reqwest::{
-    Client as ReqwestClient, Url,
-    header::{HeaderMap, HeaderValue, USER_AGENT},
-};
-
-use ogcapi_types::common::{
-    Collection, Conformance, LandingPage, Link,
-    link_rel::{CONFORMANCE, DATA, NEXT},
-};
-use ogcapi_types::features::{Feature, FeatureCollection};
-#[cfg(feature = "stac")]
-use ogcapi_types::stac::SearchParams;
-
-use crate::Error;
-
-type BoxFuture<T> = Pin<Box<dyn Future<Output = Result<T, Error>>>>;
-
-/// Async client to access OGC APIs and/or SpatioTemporal Asset Catalogs (STAC).
-///
-/// # Example
-///
-/// ```rust,ignore
-/// use ogcapi_client::Client;
-///
-/// #[tokio::main]
-/// async fn main() {
-///     let client = Client::new("https://example.com/ogc/").unwrap();
-///     let collection = client.collection("my-collection").await.unwrap();
-/// }
-/// ```
-#[derive(Clone)]
-pub struct Client {
-    pub(crate) client: ReqwestClient,
-    pub(crate) endpoint: Url,
-}
-
-impl Client {
-    /// Creates a Client for a given OGC API endpoint.
-    pub fn new(endpoint: &str) -> Result<Self, Error> {
-        let mut headers = HeaderMap::new();
-        headers.insert(USER_AGENT, HeaderValue::from_static(crate::UA_STRING));
-
-        let client = ReqwestClient::builder()
-            .default_headers(headers)
-            .build()
-            .expect("Build a client");
-
-        Self::new_with(endpoint, client)
-    }
-
-    /// Creates a Client with a custom `reqwest::Client`.
-    pub fn new_with(endpoint: &str, client: ReqwestClient) -> Result<Self, Error> {
-        let endpoint = if endpoint.ends_with('/') {
-            endpoint.parse::<Url>()?
-        } else {
-            format!("{endpoint}/").parse::<Url>()?
-        };
-
-        Ok(Self { client, endpoint })
-    }
-
-    /// Returns the landing page.
-    pub async fn root(&self) -> Result<LandingPage, Error> {
-        self.fetch::<LandingPage>(self.endpoint.as_ref()).await
-    }
-
-    /// Returns the conformance declaration.
-    pub async fn conformance(&self) -> Result<Conformance, Error> {
-        let root = self.root().await?;
-
-        #[cfg(feature = "stac")]
-        if !root.conforms_to.is_empty() {
-            return Ok(Conformance {
-                conforms_to: root.conforms_to,
-            });
-        }
-
-        if let Some(link) = root.links.iter().find(|l| l.rel == CONFORMANCE) {
-            return self.fetch::<Conformance>(&link.href).await;
-        }
-
-        Err(Error::UnknownConformance(
-            "Unable to retrieve conformance.".to_string(),
-        ))
-    }
-
-    /// Returns an async paginating iterator over collections.
-    pub async fn collections(&self) -> Result<CollectionsStream, Error> {
-        let root = self.root().await?;
-
-        if let Some(link) = root.links.iter().find(|l| l.rel == DATA) {
-            let page = self
-                .fetch::<ogcapi_types::common::Collections>(&link.href)
-                .await?;
-            Ok(CollectionsStream {
-                client: self.clone(),
-                collections: page.collections.into_iter(),
-                links: page.links,
-                pending: None,
-            })
-        } else {
-            Err(Error::ClientError(
-                "No link found with relation `data`!".to_string(),
-            ))
-        }
-    }
-
-    /// Returns a single collection by id.
-    pub async fn collection(&self, id: &str) -> Result<Collection, Error> {
-        let url = self.endpoint.join(&format!("collections/{id}"))?;
-        self.fetch::<Collection>(url.as_str()).await
-    }
-
-    /// Returns an async paginating iterator over items in a collection.
-    pub async fn items(&self, id: &str) -> Result<ItemsStream, Error> {
-        let url = self.endpoint.join(&format!("collections/{id}/items"))?;
-        let page = self.fetch::<FeatureCollection>(url.as_str()).await?;
-        Ok(ItemsStream {
-            client: self.clone(),
-            items: page.features.into_iter(),
-            links: page.links,
-            pending: None,
-        })
-    }
-
-    /// Returns an async paginating iterator over items in a collection,
-    /// filtered by the given query parameters (bbox, datetime, limit, etc.).
-    pub async fn items_with_query(
-        &self,
-        id: &str,
-        query: &ogcapi_types::features::Query,
-    ) -> Result<ItemsStream, Error> {
-        let base = self.endpoint.join(&format!("collections/{id}/items"))?;
-        let qs = serde_qs::to_string(query)?;
-        let url = if qs.is_empty() {
-            base.to_string()
-        } else {
-            format!("{base}?{qs}")
-        };
-        let page = self.fetch::<FeatureCollection>(&url).await?;
-        Ok(ItemsStream {
-            client: self.clone(),
-            items: page.features.into_iter(),
-            links: page.links,
-            pending: None,
-        })
-    }
-
-    /// Searches items with the given parameters.
-    #[cfg(feature = "stac")]
-    pub async fn search(&self, params: SearchParams) -> Result<ItemsStream, Error> {
-        let url = format!("{}search?{}", self.endpoint, serde_qs::to_string(&params)?);
-        let page = self.fetch::<FeatureCollection>(&url).await?;
-        Ok(ItemsStream {
-            client: self.clone(),
-            items: page.features.into_iter(),
-            links: page.links,
-            pending: None,
-        })
-    }
-
-    pub(crate) async fn fetch<T>(&self, url: &str) -> Result<T, Error>
-    where
-        T: serde::de::DeserializeOwned,
-    {
-        log::debug!("Fetching {url}");
-
-        self.client
-            .get(url)
-            .send()
-            .await
-            .and_then(|rsp| rsp.error_for_status())
-            .map_err(Error::RequestError)?
-            .json::<T>()
-            .await
-            .map_err(Error::RequestError)
-    }
-}
-
-/// Async paginating iterator over collections. Implements [`Stream`].
-pub struct CollectionsStream {
-    client: Client,
-    collections: <Vec<Collection> as IntoIterator>::IntoIter,
-    links: Vec<Link>,
-    pending: Option<BoxFuture<ogcapi_types::common::Collections>>,
-}
-
-impl Stream for CollectionsStream {
-    type Item = Result<Collection, Error>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.get_mut();
-
-        loop {
-            if let Some(fut) = &mut this.pending {
-                match fut.as_mut().poll(cx) {
-                    Poll::Pending => return Poll::Pending,
-                    Poll::Ready(Ok(page)) => {
-                        this.collections = page.collections.into_iter();
-                        this.links = page.links;
-                        this.pending = None;
-                    }
-                    Poll::Ready(Err(err)) => {
-                        this.pending = None;
-                        this.links = Vec::new();
-                        return Poll::Ready(Some(Err(err)));
-                    }
-                }
-            }
-
-            if let Some(value) = this.collections.next() {
-                return Poll::Ready(Some(Ok(value)));
-            }
-
-            if let Some(link) = this.links.iter().find(|l| l.rel == NEXT) {
-                let href = link.href.clone();
-                let client = this.client.clone();
-                this.pending = Some(Box::pin(async move {
-                    client
-                        .fetch::<ogcapi_types::common::Collections>(&href)
-                        .await
-                }));
-            } else {
-                return Poll::Ready(None);
-            }
-        }
-    }
-}
-
-/// Async paginating iterator over items. Implements [`Stream`].
-pub struct ItemsStream {
-    client: Client,
-    items: <Vec<Feature> as IntoIterator>::IntoIter,
-    links: Vec<Link>,
-    pending: Option<BoxFuture<FeatureCollection>>,
-}
-
-impl Stream for ItemsStream {
-    type Item = Result<Feature, Error>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.get_mut();
-
-        loop {
-            if let Some(fut) = &mut this.pending {
-                match fut.as_mut().poll(cx) {
-                    Poll::Pending => return Poll::Pending,
-                    Poll::Ready(Ok(page)) => {
-                        this.items = page.features.into_iter();
-                        this.links = page.links;
-                        this.pending = None;
-                    }
-                    Poll::Ready(Err(err)) => {
-                        this.pending = None;
-                        this.links = Vec::new();
-                        return Poll::Ready(Some(Err(err)));
-                    }
-                }
-            }
-
-            if let Some(value) = this.items.next() {
-                return Poll::Ready(Some(Ok(value)));
-            }
-
-            if let Some(link) = this.links.iter().find(|l| l.rel == NEXT) {
-                let href = link.href.clone();
-                let client = this.client.clone();
-                this.pending = Some(Box::pin(async move {
-                    client.fetch::<FeatureCollection>(&href).await
-                }));
-            } else {
-                return Poll::Ready(None);
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use futures_util::TryStreamExt;
-
-    #[tokio::test]
-    async fn conformance() {
-        let endpoint = "https://data.geo.admin.ch/api/stac/v0.9/";
-        let client = Client::new(endpoint).unwrap();
-        let conformance = client.conformance().await.unwrap();
-        println!("{conformance:#?}");
-        assert!(!conformance.conforms_to.is_empty());
-    }
-
-    #[tokio::test]
-    async fn collection() {
-        let endpoint = "https://data.geo.admin.ch/api/stac/v0.9/";
-        let client = Client::new(endpoint).unwrap();
-        let collection = client.collection("ch.swisstopo.swissalti3d").await.unwrap();
-        assert_eq!("ch.swisstopo.swissalti3d", collection.id);
-    }
-
-    #[tokio::test]
-    async fn collections() {
-        let endpoint = "https://data.geo.admin.ch/api/stac/v0.9/";
-        let client = Client::new(endpoint).unwrap();
-        let collections: Vec<_> = client
-            .collections()
-            .await
-            .unwrap()
-            .try_collect()
-            .await
-            .unwrap();
-        for c in &collections {
-            println!("{}", c.id);
-        }
-        assert!(!collections.is_empty());
-    }
-
-    #[tokio::test]
-    #[cfg(feature = "stac")]
-    async fn search() {
-        let endpoint = "https://data.geo.admin.ch/api/stac/v0.9/";
-        let client = Client::new(endpoint).unwrap();
-        let bbox = ogcapi_types::common::Bbox::from([7.4473, 46.9479, 7.4475, 46.9481]);
-        let params = ogcapi_types::stac::SearchParams::new()
-            .with_bbox(bbox)
-            .with_collections(["ch.swisstopo.swissalti3d"]);
-        let item = client
-            .search(params)
-            .await
-            .unwrap()
-            .try_next()
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(
-            Some("ch.swisstopo.swissalti3d".to_string()),
-            item.collection
-        );
-    }
-}
